Only look for address in `to` field for letters we send by DVLA api
We were getting the recipient address from the `to` field of the
notification if the letter was precompiled, and the personalisation if it
wasn't. This was because one-off letters only stored the first line of the
address in the `to` field - all other types had the full address in that
field.

We've now made a change to store the whole address in the `to` field for
one-off letters (https://github.com/alphagov/notifications-admin/pull/4655).
This means that we don't need to check how the letter was created in
order to see where the address is stored - it will always be in the `to`
field.