Remove custom sampling logic
We are getting persistent memory leaks/significantly increased memory
usage with sentry deployed. I suspect this is related to tracing
requests based on parent_sampled. For now the most important thing for
us is getting error reports. Let's make tracing dumber (don't trace even
if the parent request was traced) so that we can focus on that priority.

But let's retain the setting so that we can continue to trace in
preview+staging if we want to.