Reformat __str__ output for some errors
Rather than stringify the dict outright let's try to generate a more
human-friendly message.

The main patch here is intended to stop this breaking our custom utils
JSONFormatter code, which will call `.format()` on the log message.

When the log message contains {} it will throw an error if it can't find
a key in the format data. We're logging exceptions on basically ever
InvalidRequest because of failures to format the log message string.

This is a quick workaround that will address a lot of broken log
messages. We should still give some more thought to our utils
JSONFormatter. I think we should remove the implicit call to `.format()`
and do logging in a more common/best practice way, TBC.