Schedule letter collation twice a day; run once
Before the timezone change - and us switching celery-beat onto UTC
rather than Europe/London - we were always collating letters at 17:50
local time.

Since changing to BST this has now been happening at 18:50 BST (17:50
UTC). We'd like to make this so that we always send the data over at
17:50 in local time again.

In order to make this happen, we will schedule the task to run twice in
UTC in celery-beat. One of these runs will equate to 17:50 GMT and the
other to 17:50 BST. The task itself then needs to only process during
the correct window and abort if it's processing early or late.