Stop on-demand calculation of ft_billing data
Remove the on-demand calculations of ft_billing when the API endpoint is
hit, as this can be a set of heavy queries causing slow page loads. We
now have an hourly celery task that calculates this information for the
current day, so we can use the cached information instead.

To indicate when the data was last updated, we'll return the timestamp
for when the cache was last updated.