Add AnnualAllowance table
We have a Rates table that we use to keep track of the price of text
messages and how they change over time.

We review annual free text fragment allowances quite regularly (at least
annually) and we'd like a similar way of managing those allowances.

Previously we had the allowances hard-coded in the API and this caused
some issues when we entered a financial year that we didn't have coded
allowances for.

We will now have allowances that are considered valid in perpetuity -
until a new rate is set to override it.