Pull default annual allowances from the DB
Remove the hard-coded dict of annual allowances and use the new
DefaultAnnualAllowance model/table instead.

We don't have historical information from before 2020 so if we tried to
populate a default for 2019 or earlier this would throw an error. I
don't think we need to worry about this case, so ignoring it. Hopefully
no-one reads this commit in the future and finds out this was a poor
assumption.