Remove special case for copying 0 allowances
This command had a special check to detect whether or not a service had
a 0 free sms fragment allowance in the previous year and copies it
forward if so.

We have moved the responsibility for this into
`set_default_free_allowance_for_service` now, so the command doesn't
need to do it as well.