Adding archive endpoint for letter-attachment
This endpoint is the one called when removing an attachment from a
template.

As there are multiple versions of the template, which might be reverted
to, the attachment is kept. So this archive endpoint just does a soft
delete.

It does this by removing it from template and updating the
archived_at and archived_by_id in the letter-attachment
db row.