Split the URLs used in the document download client
We were using upload URL in the document download client for two things:
1. to upload a document
2. to display in the response when sending an email to a simulated
  recipient

We want to use the internal route where possible when the app is
deployed on ECS. When deployed on the PaaS, both
`DOCUMENT_DOWNLOAD_API_HOST` and `DOCUMENT_DOWNLOAD_API_HOST_INTERNAL`
have the same value.

This changes the document download client to use the internal url when
uploading a document and the external url for simulated email responses.