return full letter attachment object if available
this is useful, as we typically pass the template json blob straight to
template preview without modification when generating letters. If we
return the full letter attachment object (note this is just id,
metadata, etc, not the actual PDF), then that's one less thing we have
to worry about handling separately elsewhere.