Use `pytest.approx()` to compare float value
When running the tests in Postgres 15 we saw an extra decimal place
returned for the value of `percentage_under_10_seconds`. We never need
this value to be this precise (we round it in admin) and comparing
floats has limitations, so this uses `pytest.approx()`, which is
designed for floating point comparisions, in the assertion.