Add `NotificationLetterDespatch` table
This table will keep a record of all letter notifications that we have
received confirmation of despatch for. It will be populated as files are
dropped into FTP by our print provider at the same time as some other
tasks (daily sorted count and updating the main notifications table).

We will use this table to build a new fact billing table that will be
used to reconcile invoices/statements from the print provider, as they
want us to bill around despatch date rather than the date that our users
submit the letters to us.