Add facts table and generate facts for letter despatch
Set up and start populating a facts table for letter despatches so that
we can start billing DVLA based on despatch date rather than our
notification created date.

This closely follows the existing fact billing process for normal
notifications with one distinction: if we regenerate
FactBillingLetterDespatch it will also delete any records that should no
longer exist; this isn't the case for the standard FactBilling table and
those need to be manually deleted.