Start recording letter despatch data
DVLA have started to send us data in their receipt files about when a
given letter was despatched and want us to bill according to that.

As we process receipts, let's store a persistent copy of that data. In
the future we'll be able to use this to build a fact billing table that
we can use for a billing report.

We persist this in the database as the receipt files are only retained
for 14 days - this gives us an accessible on-demand record of when letters
were despatched. Having this data in a table means that we can follow a
similar fact billing generation process as we do for other reports; ie
we can just drop rows and regenerate the full facts easily.