Remove use of `select *` in a query
This is done for two reasons.

Firstly, it is bad practice to use `select *`, backed up by
https://stackoverflow.com/questions/321299/what-is-the-reason-not-to-use-select

Secondly, we've got a new database in our staging environment that
has the same columns in the notifications table, but the order of
the columns is slightly different. `select *` returns data in
the default order and this order is different between the
`notifications` and `notifications_history` tables meaning that
this query currently errors. This column ordering difference was
introduced in:
https://github.com/alphagov/notifications-aws/blob/main/scripts/remove-all-sensitive-and-personal-data/remove-all-sensitive-and-personal-data-part-2.sql