Move repeated string of notification_history column names to list
We start by moving it into a single place and storing as a list so
our code becomes a bit more DRY.

We also add a test to make sure that we remember to update this
list whenever a new column is added to `NotificationHistory`.

We did consider doing something smart like:
```
[value.name for value in column.NotificationHistory.__table__.columns.values()]
```
to automatically get the column names of the NotificationHistory
table but thought we would go for more readable, explicit and less
magic.