Create functional test seeded users endpoint
Adds an endpoint to the API, only available in development and preview
environments, which can create users for the functional tests.

These users are created against a pre-seeded organisation/service, so
that is expected to exist in the DB already. But creating users on
demand allows us to run tests interacting with that org/service in
parallel, which should lead to much faster run times for our common
journeys