Set `result_expires` to 0 in Celery config
When running Celery beat on ECS we've noticed the `backend_cleanup` task
running at 0400 each morning and being put on the default queue. This
task is a built in Celery task that cleans up expired results ("task
tombstones") from the results backend database. We don't have a results
backend, so don't need the task to run - there's nothing for it to clean up.
We can disable it by setting the value of `results_expires` to 0 to
show that task results never expire and so don't need cleaning.

It's not clear why this task is running on ECS but wasn't running on the PaaS
since we've not changed any settings when migrating.