Make fragment part of links to github lowercase
It looks like github (I assume, unless it is a change of browser
behaviour for me?) runbook links are no longer sending you to
the correct part of the page. This appears to be because the
fragment part of the link now needs to be lowercased.

You can test this by seeing that
https://github.com/alphagov/notifications-manuals/wiki/Support-Runbook#Deal-with-services-with-high-failure-rates-or-sending-sms-to-tv-numbers
does not take you to the right section but
https://github.com/alphagov/notifications-manuals/wiki/Support-Runbook#deal-with-services-with-high-failure-rates-or-sending-sms-to-tv-numbers
does.

I've searched the API for all references to our github wiki
to make sure all fragments are completely lower case.