Cleanly set up seeded user organisation
When we receive a request to set up seeded users for the functional
tests, let's set the organisation for a user in a reliable way:

1) Remove them from any organisations they're currently in, in case we
don't pass `organisation_id` (which says they shouldn't be in any org)
2) Set them up in the organisation via the `dao_` method, which will
also set any permissions accordingly.

We follow this process even if they're asking to be a member of the org
they're already in. This is mostly preparing for a future case where we
may pass permissions, and using the `dao_add_user_to_organisation` will
then cleanly ensure they have only the correct permissions.