Add celery task to delete one quarter of old notification history
This will run at 9pm (a quietish time but hopefully not going
to conflict with the other nightly tasks that start around midnight,
although if it does conflict it should still be OK).

This will start at 9pm on the day of merging and then should
run for 14 days, each time deleting the oldest quarter of data.
When it runs on the 15th day (and every day after that), it will
see there are no notifications older than 1 April 2023 and do nothing.
We could at that point, delete the celery beat configuration for it
so we stop running a task that isn't going to do anything every night.

I assume that deleting one quarter of data at a time will be OK. The
query will probably take quite a long time to run and delete a
quarter of data, probably a few hours. One risk is that the
celery work will get scaled down or restarted during this time.
That shouldn't be the end of the world as the task will just run
again tomorrow. Deleting notification_history isn't some urgent thing
that must happen every night, if it gets delayed by a day that is
fine. If we do start seeing issues with these queries taking
several hours and erroring for some reason, we can take a similar
approach to other mass deletion functions we have which run
multiple queries, each one deleting 10,000 records at a time. A bit
more complex python is all that is needed.

In july of 2024, we will be able to update the retention limit to
April 2024 if we want so that this process will happen again
for next years deletion.