always connect to the test database
previously we were relying on the database name being passed through in
the SQLALCHEMY_DATABASE_URI variable (or being set by the default in the
config). this doesn't play well with running the tests in the same
docker container we use to run the app locally, so update the code so
that the database name is hardcoded in the db fixture. I don't think
this is a huge deal as we're modifying the database name (to create a
separate db for each pytest worker) at this point anyway