Allow inbound sms if sending sms permission is off
We recently encountered the case where a service set up inbound
sms and then turned off permission to send sms. This means it has
the `inbound-sms` permission but not the `sms` permission.

When you turn off SMS in the admin app for your service, it hides
all the SMS related settings, including inbound sms. This is
probably not ideal. We should consider changing this in the future
so that the admin interface allows you to configure inbound and
outbound sms separately.

We also had some discussion about whether turning off outbound sms
should also turn off inbound sms but decided this was not the best
for the users and we should allow them the flexibility.

Given that we want services to still be able to receive inbound
sms, even if they have outbound sms turned off, it requires this
change to the permissions required to receive inbound sms.

There was no context in the commit that added this behaviour back
in 2017 that this is a very intentional decision that should not
be reversed.
https://github.com/alphagov/notifications-api/commit/5a82fe0a70bb86bd0e3df7c16498af66e9888e24