Add welsh_subject, welsh_content and languages columns to templates
table.

Also add a new enum: template_language_options that specifies supported
language options for bilingual letter templates.

NOTE: automatically created migration for this did not work and
I had to substantially modify it.
Automatic command tried to create enum at the same time it was
adding a column relying on that enum for its default.
This threw an error.
After some searching, I found this stackoverflow answer:
https://stackoverflow.com/a/68800087

I did what it recommended, that is, I split the operations up a bit:
I first defined the new enum, then created it, and only then added
the column that relies on that enum. That worked a treat.