Fix case where oldest notification_history record is start of quarter
If the oldest notification is at 1 Jan 2023, we still need to
delete from 1 Jan 2023 up to 30 March 2023 23:59:59, for example.

Otherwise, this would likely mean that our script would start by
deleting the first quarter of data and then get no further if there
was a notification created at exactly the very beginning of the
next quarter.