populate normalised_service_name from email_from
when email_from is updated (or provided during service creation), update
normalised_service_name. if normalised_service_name is also provided,
use that instead.

I tried to use the schema but had issues with the update function where
it proved hard to find out what the user had changed and what was
existing values - since this code is only temporary its fine to just
leave it in the rest functions i think