Handle BST for delete_oldest_quarter_of_unneeded_notification_history
Our previous code when run multiple times would have deleted
anything older than 1 April 2023. But we actually need to keep
everything from 11pm 31 March 2023 because our billing is done in
BST quarters.

This commit makes the adapations to make sure that we keep
the full BST quarters of data that we need, not the UTC ones. It
adds a fair bit of complexity to the code which I've tried to make
as readable as possible to combat this.

I've gone for the do as much as possible in BST approach and convert
to UTC at the final step, however it would have been equally valid
to just do everything in UTC throughout, but would have added
code complexity to the calculations when iterating through the
quarters to find which one to delete.