Revert "Bump gunicorn to 21.2.0"
This reverts the changes introduced by commit
6d80b03c577bb47cd09ac3e49a4b30284179832b.

In a recent load test we identified a huge drop-off in performance that
has been traced back to this dependency bump. Until we can investigate
further and work out exactly why - and potentially get it fixed up
stream - we will go back to our previously-pinned version.