Fix implementation of Notification.substitute_status
We were returning a mix of types: a string in some cases and a list of
strings in others. We now consistently return a list of strings, which
is what usage of the method expects anyway, as we pass the return value
into a SQLAlchemy `in` clause, which will iterate over the return value.
If it's a string, we end up checking if a notification status is in (eg)
'delivered', which checks against 'd,' 'e', 'l', ... We should instead
checks against ['delivered'].