Add index concurrently on created_at to notification_history table
We want to add this index because at the moment it the table is
too big to query based on created_at (we do have some indexes
on this table already but those are multi column indexes so are
not often able to be used when querying just on created_at).
A specific example for where this index will be used is in
https://github.com/alphagov/notifications-api/pull/3860.

We choose to create our index concurrently. This is backed by
https://medium.com/paypal-tech/postgresql-at-scale-database-schema-changes-without-downtime-20d3749ed680#04b4.

Importantly it says:
This lower lock level allows reads and writes to continue against the
table while the index is built.

You can also read the postgres documentation on it at:
https://www.postgresql.org/docs/current/sql-createindex.html#SQL-CREATEINDEX-CONCURRENTLY

Creating an index concurrently needs to be done outside of
a transaction. We can use an alembic feature to support the migration
being done outside of a transaction:
https://alembic.sqlalchemy.org/en/latest/api/runtime.html#alembic.runtime.migration.MigrationContext.autocommit_block

Because it is done outside of a transaction, the sql should
run very quickly and the migration should suceed, leaving the
index to be built concurrently happening behind the scenes.

I will need to manually check that
the index is eventually created successfully. If it fails, then it
will need to be removed manually, as per the postgres docs.

I'm not too worried about performance by adding this index. We
don't care too much about insertion speed for the notification_history
table, which an index might affect. It should however speed up
our ability to query notification_history for particular
queries.