run upgrades in autocommit mode to remove deadlock
we ran into deadlocks when running this script on production. adding a
constraint acquires an access-exclusive lock. This is fine on its own,
to add a not-null constraint is a relatively quick operation for a table
of this size (27k services, 100k services_history). however, because the
query was trying to acquire locks to both services and services_history,
it ran in to a deadlock with another process where that process was also
trying to acquire locks on both tables - each process acquired one of
the locks, and was stuck waiting for the other to be released, so a
deadlock was identified and the migration failed.

even if the other process wasn't doing anything big and dramatic,
nothing can acquire a lock alongside an access-exclusive lock.

normally alembic wraps the entire upgrade script in a transaction (so
that if there's an error, it can rollback safely to the previous version
and ensure that the db state is consistent). but we can manually enter
an autocommit block such that every statement commits separately outside
of a transaction.

this has already been applied succesfully without incident to preview
and staging so there's no action to do there.