set transaction_per_migration=True for alembic
alembic wraps any migrations in a transaction - this is critical so if a
single line fails, alembic can roll back any changes to the last known
good state. However, by default it will wrap all migrations to be run in
a single transaction.

This flag changes it to have one transaction per migration file,
meaning that if we're deploying two migration files at once, and the
second one has an error, it won't rollback the first one. It also,
critically, will ensure any locks required for the first migration won't
be held for the duration of the second one