lint alembic migrations with squawk
squawk can take a sql file and lint it for common errors, such as
renaming a column in an unsafe way or running a command that might
acquire a wide-reaching long-lasting lock.

this commit wraps squawk in a little bash script that gets the alembic
upgrades output as sql

there doesn't appear to be an obvious way to "accept" a squawk warning
if we know it's okay (for example, dropping a column we know is no
longer written to or read from), so we can't immediately integrate this
in to our CI pipeline or pre-commit hooks unfortunately. but we can
consider adding this as a comment on pull requests or similar