Delete rows from notification_history in 50,000 chunks
This is because the query is currently taking over 20 minutes
which is longer than our statement timeout limit and so is being
cancelled.

We split it into smaller chunks so that we don't run into this
problem. 50,000 has been chosen fairly arbirtrarily and is used
as a query_limit elsewhere in our codebase.

We commit after each one so we commit progress regularly rather
than just at the end.

We choose synchronize_session False as we do not need to access
those notifications again that have been deleted.
https://docs.sqlalchemy.org/en/14/orm/session_basics.html#selecting-a-synchronization-strategy