Don't run notification_history purge task in staging
This will enable us to retain about 1 billion notifications in
our staging notification_history table, which is roughly the same
number that will be in production after this task has run against
prod. This is important for our testing this week of migrating
Postgres which will benefit from having staging look similar
to production in size.

We did also consider taking a fresh copy of production, redacting
the personal data and attaching it to the staging apps, but decided
that was too much work and this is a quicker fix.