Use `days` instead of `hours` for scheduled jobs
This is the intention we are trying to get across – that jobs can be
scheduled  up to 4 days in the future, and we don’t have to worry about
what magic number 96 means.

From experimenting it seems like `hours=24` and `days=1` are equivalent
when it comes to crossing a daylight-savings-time boundary. They both
advance the wall clock by 24 hours, rather than by 24 elapsed hours.