Delete notification_history in hour chunks
This changes the approach in
delete_oldest_quarter_of_unneeded_notification_history
which was too slow because it couldn't be run in parralell on
multiple workers.

Instead, we queue up lots of small tasks that can be run in parraell
which will each delete only a single hour of data.

I've tested some of these queries manually on production and for
the 4 different 1 hour slots I tried, the queries took less than
10 seconds each.

The plan is to kick this task off once manually by logging
into production and creating a task (rather than scheduling a task
that we only want to run once which is a bit hard to do with cron).