add new email sender columns to service and history table
the `custom_email_sender_name` will be set by users with manage service
permission on the front-end, to reflect a name they'd like their emails
to appear from.

the `email_sender_local_part` will be set by lowercasing that value, and
replacing spaces and punctuation with the period character. if the
service hasn't set `custom_email_sender_name` then the local part should
be set to be equal to normalised_service_name. eventually this column
will have a not-null constraint. (we could also add a db constraint to
check the equivalence of normalised_service_name too if we want)

On max data length:

The RFC3696[^1] states that local parts may be up to 64 characters.
However, in practice, we have seen notify services succesfully sending
emails with local parts longer than that - the key limitation is that
the entire email address must be under 320 characters.

The entire from address is `"Service Name"
<service.name@notifications.service.gov.uk>` (and this may change in the
future if we add custom from domains).

This includes 34 characters for the prod notify domain, @, angle
brackets, double quotes etc.

Set the limit to 255 characters for both columns just to be flexible and
comparable to existing name/normalised_service_name columns. We'll
validate that the produced email address is under 320 characters in the
admin app.

[^1] https://datatracker.ietf.org/doc/html/rfc3696#:~:text=In%20addition%20to%20restrictions,they%20are%20rarely%0A%20%20%20encountered.