Raise sentry error for slow text messages
Although we have nice graphs for monitoring slow text message delivery,
they currently require proactive manual monitoring to spot slow delivery
issues across the board.

Let's try to be a bit more proactive on this. We can check whether >10%
or more of text messages have been slow (>5 minutes) in the last 15
minutes, and raise an error if slow. This should let us get a jump on
investigating problems here.

10% is a _bit_ of a finger in the air from looking at the recent slow
delivery graphs and trying to work out a reasonable point to be alerted
at that is a balance between not being too noisy, and detecting levels
of user impact that aren't tolerable.