Reduce number of slow sms delivery errors
Let's only log an error (and thereby raise a Sentry error) if we've seen
delivery be slow for >10% of text messages for the last 20 minutes
consecutively. This should help increase the signal of the error,
indicating that every time it happens we actually need to take a look
and investigate properly. It's been a bit noisy and flappy so far.

We require 5 consecutive failures. The task runs every minute, so this
essentially mean for 5 minutes consistently (and checks ~20 minutes of
activity altogether).